+++ app/src/main/java/com/oriol/oompasmanager/presentation/ompalist/OmpaListViewModel.kt
@@ -1,16 +1,17 @@
-package com.oriol.oompasmanager.view.ui
+package com.oriol.oompasmanager.presentation.ompalist
 
 import androidx.lifecycle.MutableLiveData
-import com.oriol.oompasmanager.model.ApiRepository
-import com.oriol.oompasmanager.model.entities.ResultsItem
+import com.oriol.oompasmanager.domain.model.ResultsItem
+import com.oriol.oompasmanager.domain.usecase.GetOmpaListUseCase
+import com.oriol.oompasmanager.presentation.BaseViewModel
 
-class OmpaListViewModel: BaseViewModel() {
+class OmpaListViewModel constructor(private val getOmpaListUseCase: GetOmpaListUseCase): BaseViewModel() {
 
     val ompaListLive = MutableLiveData<List<ResultsItem?>>()
 
     fun fetchOmpaList() {
         dataLoading.value = true
-        ApiRepository.getInstance().getOmpasList { isSuccess, response ->
+        getOmpaListUseCase.getOmpasList { isSuccess, response ->
             dataLoading.value = false
             if (isSuccess) {
                 ompaListLive.value =  response?.results
