+++ app/src/main/res/layout/fragment_ompa_list.xml
@@ -5,7 +5,7 @@
         <import type="android.view.View" />
         <variable
                 name="viewmodel"
-                type="com.oriol.oompasmanager.view.ui.OmpaListViewModel" />
+                type="com.oriol.oompasmanager.presentation.ompalist.OmpaListViewModel" />
     </data>
 
     <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
